keep numpy out of the exposure time
Live on the 2026-07-27 twilight: eight

    bus: serialization issue, won't work on remote buses
    TypeError: Encoding objects of type numpy.float64 is unsupported

per frame - the expose request plus every get_metadata call behind it,
each carrying exptime=np.float64(48.734173006548).

np.exp() in the sky model returns a numpy scalar, so the rate is one, and
the calculator's closing partial step (ideal_counts - counts) / rate hands
that type to the exposure time, the ImageRequest, and from there to every
instrument the frame collects headers from. Same family as the sky level
cast in f4b5ff8, one layer up.

The rate is now cast where it leaves the model, and _take_image casts again
at the ImageRequest - whatever a caller passes, the request that crosses the
bus carries plain floats.

Harmless on a single-host bus, where Bus._push only encodes to check and
delivers the object anyway; on a split deployment it would have taken the
exposure down. The fake camera now encodes every request with msgspec, so
any test that takes a frame fails on a numpy scalar, and the calculator's
return type is pinned directly.

+++ src/chimera_skyflat/controllers/autoskyflat.py
@@ -176,7 +176,9 @@ class AutoSkyFlat(ChimeraObject, IAutoSkyFlat):
         self._set_filter(filter_id)
 
         imrequest = ImageRequest(
-            exptime=exptime,
+            # plain float: msgspec cannot encode a numpy scalar, and this
+            # request is copied into every metadata call the frame makes
+            exptime=float(exptime),
             frames=1,
             shutter=Shutter.OPEN,
             filename=os.path.basename(ImageUtil.make_filename("skyflat-$DATE-$TIME")),
@@ -624,8 +626,13 @@ class AutoSkyFlat(ChimeraObject, IAutoSkyFlat):
 
             while exposure_time <= exptime_max:
                 # the sun keeps moving during the exposure, so integrate the
-                # rate forward instead of freezing it at the start
-                rate = self._sky_rate(sun_alt + sun_rate * exposure_time, model_gain)
+                # rate forward instead of freezing it at the start. float():
+                # np.exp() makes the rate a numpy scalar, and msgspec cannot
+                # encode one, so letting it reach the ImageRequest breaks
+                # every bus message the exposure touches
+                rate = float(
+                    self._sky_rate(sun_alt + sun_rate * exposure_time, model_gain)
+                )
                 if counts + rate * increment >= ideal_counts:
                     # finish on a partial step: rounding the exposure down to
                     # a whole increment costs up to exptime_increment of sky,
